ipex: support xpu streaming
