+++ requirements-build.txt
@@ -1,5 +1,5 @@
 # Should be mirrored in pyproject.toml
 packaging
 setuptools>=49.4.0
-torch==2.3.1
+torch
 wheel
