+++ requirements-cuda.txt
@@ -1,2 +1,4 @@
+torch==2.3.1
+torchvision~=0.18.1
 onnxruntime-gpu~=1.18.0
 onnxruntime-genai-cuda~=0.3.0rc2
