+++ requirements-directml.txt
@@ -1,2 +1,4 @@
+torch==2.3.1
+torchvision~=0.18.1
 onnxruntime-directml~=1.18.0
 onnxruntime-genai-directml~=0.3.0
