+++ requirements-xpu.txt
@@ -1,0 +1,4 @@
+torch==2.1.0
+torchvision
+trl
+transformers~=4.42.3
